+++ src/test/utils.test.ts
@@ -1,5 +1,5 @@
 import * as assert from 'assert';
-import { capitalizeComponentName, removeNonWordCharacters } from '../utils';
+import { capitalizeComponentName, removeNonWordCharacters, truncateType } from '../utils';
 
 suite('removeNonWordCharacters', function () {
   test('should remove non-word characters from the string', function () {
@@ -43,6 +43,36 @@ suite('capitalizeComponentName', function () {
     const input = 'Button';
     const expected = 'Button';
     const result = capitalizeComponentName(input);
+    assert.strictEqual(result, expected);
+  });
+});
+
+suite('truncateType', function () {
+  test('should return the same string if its length is less than or equal to 500', function () {
+    const type = 'string';
+    const result = truncateType(type);
+    const expected = type;
+    assert.strictEqual(result, expected);
+  });
+
+  test('should return "any" if the string length is more than 500', function () {
+    const type = 'a'.repeat(501);
+    const result = truncateType(type);
+    const expected = 'any';
+    assert.strictEqual(result, expected);
+  });
+
+  test('should return the same string if its length is exactly 500', function () {
+    const type = 'a'.repeat(500);
+    const result = truncateType(type);
+    const expected = type;
+    assert.strictEqual(result, expected);
+  });
+
+  test('should return an empty string if the string has length 0', function () {
+    const type = '';
+    const result = truncateType(type);
+    const expected = type;
     assert.strictEqual(result, expected);
   });
 });